Backend/requirements.txt
asgiref==3.8.1
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.7
Django==5.0.6
django-allauth==0.63.3
idna==3.7
mysqlclient==2.2.4
pycparser==2.22
PyJWT==2.8.0
requests==2.32.3
sqlparse==0.5.0
urllib3==2.2.1
django-debug-toolbar
django-cors-headers
